% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workFlows.R
\name{allNodesTable}
\alias{allNodesTable}
\title{Helper function that takes the result from the \code{\link{nodes}} function,
 which is a named list of parameter tables (from processing or consensus
 workflow), and puts it all in a single table with the names of the nodes as
 an extra column}
\usage{
allNodesTable(nodesList)
}
\arguments{
\item{nodesList}{named list of tables of workflow (node) parameters. Intended
as input here is the output from the \code{\link{nodes}} function}
}
\value{
data.frame, a large table of all node parameters
}
\description{
Helper function that takes the result from the \code{\link{nodes}} function,
 which is a named list of parameter tables (from processing or consensus
 workflow), and puts it all in a single table with the names of the nodes as
 an extra column
}
