% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDefinition.R
\name{analysisDefinition}
\alias{analysisDefinition}
\title{function that gets the first element of the AnalysisDefinitionXML
 column from the AnalysisDefinition table in a .pdResult file}
\usage{
analysisDefinition(db)
}
\arguments{
\item{db}{database access 'handle' pointing to a .pdResult file}
}
\value{
a named tree like list that contains the info like file names,
 study factors, correction factors, etc etc
}
\description{
function that gets the first element of the AnalysisDefinitionXML
 column from the AnalysisDefinition table in a .pdResult file
}
