% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{blobEstimateTypes}
\alias{blobEstimateTypes}
\title{function that attempts to assign a type to the blob (raw) lengths as found
 by dtermineBlobLengths}
\usage{
blobEstimateTypes(
  blobLengths,
  minimumNumber,
  numberOfGroups = minimumNumber,
  ratioNumberOfGroups = numberOfGroups - 1
)
}
\arguments{
\item{blobLengths}{the actual lengths (number of bytes) of the elements we
wish to assign types to. Can be 1 or more lengths}

\item{minimumNumber}{this defines the minimum number of columns a
blob/raw type column should be split into. In TMT10plex experiments, the
minimumNumber will usually be 10, becauseyou have 10 channels/abundances}

\item{numberOfGroups}{this defines how many 'groups' are present in the data.
Taking Abundances as an example: Proteone Discoverer has both the original
columns (say Abundances_1 through Abundances_2), but also columns where the
abundances, that 'belong' together, are eg averaged or some other
(statistical) measure  is calculated over a number of columns. You may have
eg 10 'Abundance channels' which are 5 samples total, each in duplo. This
means that some columns in the resulting table will need to be split in 10
different columns (the original 'Abundances') while 'grouped' columns should
be split into 5 different columns (eg the calculated means or variations of
the 'abundances' columns). Note that although not enforced by the code, the
numberOfGroups should always be equal or less than the  minimumNumber
parameter. Default value = minimumNumber}

\item{ratioNumberOfGroups}{when ratios between groups are calculated we get
columns (ratio columns) that need to be split into numberOfGroups - 1
(which is the efault value)}
}
\value{
a data.frame with columns what (type) and minimumSize (number of
 variables in the blob)
}
\description{
function that attempts to assign a type to the blob (raw) lengths as found
 by dtermineBlobLengths
}
\note{
this function works with single numbers and multiple numbers

this function does not deal properly with specials, their types/
 translations are resolved in a different way

there are two ways to see potential problems with the type assignments:
 the columns may contain NA values

internal function
}
