% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{blobLength}
\alias{blobLength}
\title{attempts to determine the length (in bytes) of the individual elements of a
 blob-type column of a data.frame. It should (!) return an integer value of
 course (as all elements are supposed to have the same length). Also: if all
 elements of the column are NA, the the result will be NaN}
\usage{
blobLength(blobList)
}
\arguments{
\item{blobList}{one column of a data.frame (as a list) of blob (raw) element
type elements}
}
\value{
the length of the elements in the data.frame (or list) column. Again:
 this should be an integer
}
\description{
attempts to determine the length (in bytes) of the individual elements of a
 blob-type column of a data.frame. It should (!) return an integer value of
 course (as all elements are supposed to have the same length). Also: if all
 elements of the column are NA, the the result will be NaN
}
\note{
meant for use in debugging problems
}
