% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{calcAllIFIs}
\alias{calcAllIFIs}
\title{Wrapper function that uses \code{\link{tmt11Channels}} to calculate the
 IFI's for a set of (knock out) protein channels}
\usage{
calcAllIFIs(
  db,
  proteinsKnockedOut = knockOutProteins()$short[knockOutProteins()$knockout],
  accession = NA,
  groups = tmt11Channels(),
  joined = TRUE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{proteinsKnockedOut}{character vector that specifies the (knock out)
protein channels for which the IFI's are to be calculated}

\item{accession}{single element character vector specifying the accession
of the protein whose abundances are to be used for the IFI calculation}

\item{groups}{usually either tmt10Channels() or tmt11Channels: data.frame
that specifies which (abundance) column belongs to which knock out group}

\item{joined}{defines the type of output: if TRUE then a single data.frame
with all IFI's for all (knock out) proteins is generated. Otherwise a list
of data.frame's is generated for all proteins separately}
}
\value{
a data.frame with two columns: one with the (short) name of the
 (selected) proteins  and one with the calculated values (named IFI) or a
 list of data.frame's with the same structure
}
\description{
Wrapper function that uses \code{\link{tmt11Channels}} to calculate the
 IFI's for a set of (knock out) protein channels
}
