% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{calcData}
\alias{calcData}
\title{helper function to calculate a row-wise function (like mean, median etc)
 across a data.frame}
\usage{
calcData(
  data,
  setNAZero = NA,
  removeNAs = FALSE,
  keepData = FALSE,
  calcName = "median",
  calcFunc = stats::median,
  ...
)
}
\arguments{
\item{data}{the data.frame. Note that all rows and columns are used, so
selection, filtering, etc should be done beforehand}

\item{setNAZero}{default = NA, when NA this is ignored. Otherwise all cells
containing NA will be set to the value of setNAZero. When removeNAs = TRUE,
this parameter is ignored}

\item{removeNAs}{default = FALSE, if TRUE all rows containing NA's will be
removed via na.omit()}

\item{keepData}{if TRUE, then the original data is returned also}

\item{calcName}{name of the column with the calculated values in it}

\item{calcFunc}{function to be applied row-wise across the data.frame}

\item{...}{serves to pass on "extra" arguments on to the calcFunc function,
eg na.rm = TRUE in case of calcFunc = mean}
}
\value{
a data.frame with the calculated values as the only column or with
 the calculated values as a mew column
}
\description{
helper function to calculate a row-wise function (like mean, median etc)
 across a data.frame
}
