% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{calcIFIs}
\alias{calcIFIs}
\title{function to calculate the IFI (interference free index) of a protein
 entry in the protein table of a pdResult files. Note this can only be
 calculated on the knockout proteins in the TKO control sample: see
 \code{\link{tmt10Channels}} or \code{\link{tmt11Channels}} for the eligible
 proteins}
\usage{
calcIFIs(
  db,
  selected = "His4",
  accession = knockOutProteins()$Accession[knockOutProteins()$short == selected],
  columns = "Abundances",
  groups = tmt11Channels(),
  IFIName = "IFI",
  calcFunc = mean,
  calcName = "mean",
  na.rm = TRUE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{selected}{(short) name of the selected protein}

\item{accession}{uniprot accession code of the selected protein. If parameter
"selected" is one of the short names in \code{\link{knockOutProteins}} then
doesn't need to be specified. Note that the accession does not need to be
one of the accessions of the knockout proteins}

\item{columns}{usually this will be "Abundances". It allows the selection
of the correct (raw) columns as they come out of dfTransformRaws(), eg
Abunances_1, Abundances_2, etc}

\item{groups}{usually either tmt10Channels() or tmt11Channels: data.frame
that specifies which (abundance) column belongs to which knock out group.
Note that the 'selected' argument should be in groups}

\item{IFIName}{specifies the name to give to the calculated values, usually
"IFI"}

\item{calcFunc}{function to be applied row-wise across the data.frame. Used
in the calculation of the IFI values. Default = mean}

\item{calcName}{name of the column with the calculated values in it, used
in the related function calcData()}

\item{na.rm}{default = TRUE. This specifies that NA's should be removed when
using eg mean, median, etc}
}
\value{
a data.frame with two columns: one with the (short) name of the
 (selected) protein  and one with the calculated values (named IFI)
}
\description{
function to calculate the IFI (interference free index) of a protein
 entry in the protein table of a pdResult files. Note this can only be
 calculated on the knockout proteins in the TKO control sample: see
 \code{\link{tmt10Channels}} or \code{\link{tmt11Channels}} for the eligible
 proteins
}
