% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\docType{data}
\name{columnSpecials}
\alias{columnSpecials}
\title{specials are not numeric or integer, but have chunks of a certain size}
\format{
data.frame with columns 'names' and 'size'
}
\usage{
columnSpecials
}
\description{
specials are not numeric or integer, but have chunks of a certain size
}
\note{
each chunk exists of two bytes, first one is logical (boolean):
zero = FALSE, otherwise TRUE. Second byte = also logical: determines if
value is NA (1) or not (0)
}
\keyword{datasets}
