% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{convertRawInteger}
\alias{convertRawInteger}
\title{Converts a raw vector into its integer counterpart(s)}
\usage{
convertRawInteger(rawVector, minimumSize = 1)
}
\arguments{
\item{rawVector}{A vector of type 'raw' (blob)}

\item{minimumSize}{A number (integer) indicating how many numbers (numeric)
are in the rawVector. Default = 1. This catches potential
rawVectors that are empty/NA}
}
\value{
a list of numbers (integer) or NA (if rawVector is empty)
}
\description{
Converts a raw vector into its integer counterpart(s)
}
\note{
integer vectors are 5 bytes long, the first 4 are the actual
      number, if the last byte (5) == 0 then the result is NA
}
