% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{convertRawNumeric}
\alias{convertRawNumeric}
\title{Converts a raw vector into its numeric counterpart(s)}
\usage{
convertRawNumeric(rawVector, minimumSize = 1)
}
\arguments{
\item{rawVector}{A vector of type 'raw' (blob)}

\item{minimumSize}{A number (integer) indicating how many numbers (numeric)
are in the rawVector. Default = 1. This catches potential
rawVectors that are empty/NA}
}
\value{
a list of numbers (numeric) or NA (if rawVector is empty)
}
\description{
Converts a raw vector into its numeric counterpart(s)
}
\note{
numeric vectors are 9 bytes long, the first 8 are the actual
      number, if the last byte (9) == 0 then the result is NA

internal function
}
