% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{convertRawSpecial}
\alias{convertRawSpecial}
\title{Converts a raw vector into its boolean counterpart(s)}
\usage{
convertRawSpecial(rawVector, specialSize)
}
\arguments{
\item{rawVector}{A vector of type 'raw' (blob)}

\item{specialSize}{length of the result, ie the number of columns in the raw.
This is relevant for the so called specials. It indicates
how many booleans are in the rawVector}
}
\value{
a list of booleans or NA (if rawVector is empty)
}
\description{
Converts a raw vector into its boolean counterpart(s)
}
\note{
boolean vectors are 5 bytes long, the first 4 are the actual
      boolean (anything but 0 is TRUE), if the last byte (5) == 0 then
      the result is NA

internal function
}
