% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workFlows.R
\name{createDiagrammeRString}
\alias{createDiagrammeRString}
\title{function to create a DiagrammeR string that can be used by
 DiagrammeR::grViz() to plot a visual representation of the workflow}
\usage{
createDiagrammeRString(
  nodesTable,
  showBelow = TRUE,
  returnString = TRUE,
  hideDoubleParents = data.frame(name = c("Precursor Ions Quantifier", "Feature Mapper",
    "Reporter Ions Quantifier", "Protein Marker", "Peptide in Protein Annotation",
    "Modification Sites", "Peptide Isoform Grouper"), parent = c("last", "first", "last",
    "first", "first", "last", "first"))
)
}
\arguments{
\item{nodesTable}{output from the nodeTable function. Columns that
need to be present are node, name & parent}

\item{showBelow}{boolean, default = TRUE. Set to FALSE when troubleshooting.
Note that if set to FALSE, the parameter returnString will be ignored
It is not recommended to depend on this parameter, as it will probably be
removed in a newer version of the package}

\item{returnString}{default = TRUE. Set to FALSE when troubleshooting. Note
that the parameter showBelow makes it so that this parameter is ignored.
It is not recommended to depend on this parameter, as it will probably be
removed in a newer version of the package}

\item{hideDoubleParents}{either NA (ignored) or a data.frame specifying what
tp do in case of multiple parents. The data.frame should have the columns
name and parent. The parent column should specify
which parent to use ('first' or 'last') for connections}
}
\value{
character vector that can be passed on to DiagrammeR::grViz()
}
\description{
function to create a DiagrammeR string that can be used by
 DiagrammeR::grViz() to plot a visual representation of the workflow
}
\note{
during development it was noticed that some elements (nodes in the
 diagram) have more than one parent which is not seen in the proteome
 discoverer software of Thermo Scientific. The default data.frame 'corrects'
 known multiple parent nodes. If the parameter hideDoubleParents is
 set to NA, then the double parent connections are drawn.

an example of it's use:
(workflowInfo(db))$nodeInfo$Consensus %>%
 nodeTable() %>%
  createDiagrammeRString() %>%
   grViz()
}
