% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{db_getBlobs}
\alias{db_getBlobs}
\title{detemines which columns in a database table are of the blob (raw) type}
\usage{
db_getBlobs(db, tableName)
}
\arguments{
\item{db}{database access 'handle'}

\item{tableName}{used to pass on the name of the table containing the data}
}
\value{
a datafrane with two columns: name = columname) and type  (which
 should always be 'blob')
}
\description{
detemines which columns in a database table are of the blob (raw) type
}
\note{
this is essentially a wrap around function for db_columnInfo

meant for use in debugging problems

development function, not meant for release/export
}
