% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbClose}
\alias{dbClose}
\title{Wrapper around pool::pooClose():  closes an open database
(normally opened earlier via eg db_open())}
\usage{
dbClose(db)
}
\arguments{
\item{db}{database access 'handle' to be closed}
}
\description{
Wrapper around pool::pooClose():  closes an open database
(normally opened earlier via eg db_open())
}
