% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetAnnotatedProteins}
\alias{dbGetAnnotatedProteins}
\title{Function to get the UniqueSequenceID's for proteins which are in an protein
 annotation group. Essentially does the reverse of
 \code{\link{dbGetProteinAnnotationGroupIDs}}. The output of this function
 can serve as the input for \code{\link{dbGetProteins}}}
\usage{
dbGetAnnotatedProteins(db, proteinAnnotationGroupIDs, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{proteinAnnotationGroupIDs}{the protein annotation group ID's for which
to get the UniqueSequenceID's}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame or a character vector (SQL)
}
\description{
Function to get the UniqueSequenceID's for proteins which are in an protein
 annotation group. Essentially does the reverse of
 \code{\link{dbGetProteinAnnotationGroupIDs}}. The output of this function
 can serve as the input for \code{\link{dbGetProteins}}
}
