% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetAnnotationGroups}
\alias{dbGetAnnotationGroups}
\title{Function to get the info for (protein) annotation groups. Takes eg
 \code{\link{dbGetProteinAnnotationGroupIDs}} as input}
\usage{
dbGetAnnotationGroups(
  db,
  ProteinAnnotationGroupIDs = NA,
  columnNames = NA,
  SQL = FALSE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{ProteinAnnotationGroupIDs}{the protein annotation group ID's for which
to get information}

\item{columnNames}{allows the selection of columns to take from the table,
default = NA (all columns)}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame or a character vector (SQL)
}
\description{
Function to get the info for (protein) annotation groups. Takes eg
 \code{\link{dbGetProteinAnnotationGroupIDs}} as input
}
