% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetConsensusIDs}
\alias{dbGetConsensusIDs}
\title{get the ConsensusID's from (a set of) PeptideGroupIDs}
\usage{
dbGetConsensusIDs(db, PeptideGroupIDs, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{PeptideGroupIDs}{the PeptideGroupIDs usually come from the
TargetPeptideGroups Table. This can be in numeric or character vector format}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from the
 TargetPeptideGroupsConsensusFeatures table or a character string specifying
 a SQL query
}
\description{
get the ConsensusID's from (a set of) PeptideGroupIDs
}
