% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetMassSpectrumItems}
\alias{dbGetMassSpectrumItems}
\title{get the MassSpectrumItems info from (a set of) PeptideID's}
\usage{
dbGetMassSpectrumItems(db, peptideID, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{peptideID}{the PeptideID's usually come from the
PSMS table Table. This can be in numeric/character/data.frame format}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from the MassSpectrumItems
 table or a character string specifying an SQL query
}
\description{
get the MassSpectrumItems info from (a set of) PeptideID's
}
