% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetModificationPeptideIDs}
\alias{dbGetModificationPeptideIDs}
\title{Function to get the peptideID's 'belonging' to a modification site}
\usage{
dbGetModificationPeptideIDs(db, modificationIDs, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{modificationIDs}{the modification site identifiers to get from
the ModificationSites table. This should be the 'Id' field of a modifciation
table row}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame#'}
}
\value{
a data.frame or a character vector (SQL)
}
\description{
Function to get the peptideID's 'belonging' to a modification site
}
