% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetModificationsSitesIDs}
\alias{dbGetModificationsSitesIDs}
\title{function to get the modificationSite ID's from (a set of) proteinUniqueID's}
\usage{
dbGetModificationsSitesIDs(db, proteinUniqueIDs, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{proteinUniqueIDs}{the protein identifier for which the modificationSite
ID's are to be fetched. This is a vector of one or more integer64 (package:
bit64 ) values. In protein tables this is the UniqueSequenceUD column}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing the requested data from the
 TargetProteinsModificationSites table or a character string specifying an
 SQL query
}
\description{
function to get the modificationSite ID's from (a set of) proteinUniqueID's
}
\note{
the data from modificationSitesUd's in the result can be used to query
 the ModificationSites table via \code{\link{dbGetModificationsTable}}
}
