% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetModificationsTable}
\alias{dbGetModificationsTable}
\title{function to get data from the ModificationSides table using the
 modificiationSiteId's}
\usage{
dbGetModificationsTable(
  db,
  modificatonSitesIDs,
  columnNames = NA,
  sortOrder = NA,
  SQL = FALSE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{modificatonSitesIDs}{the modification site identifiers to get from
the ModificationSites table}

\item{columnNames}{allows the selection of columns to take from the table,
default = NA (all columns)}

\item{sortOrder}{allows for sorting of the selected columns,
default = NA, (no sorting). Other valid values are a single character
string ("ASC" or "DESC") or a character vector of the same length as the
columnNames vector containing a series of "ASC" or "DESC"}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame or a character vector (SQL)
}
\description{
function to get data from the ModificationSides table using the
 modificiationSiteId's
}
\note{
the easiest way to get the modificationSitesIDs is via the
 \code{\link{dbGetModificationsSitesIDs}} function
}
