% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetPeptideIDs}
\alias{dbGetPeptideIDs}
\title{get the peptideID's from (a set of) proteinGroupIDs}
\usage{
dbGetPeptideIDs(db, proteinGroupIDs, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{proteinGroupIDs}{the proteinGroupIDs usually come from the
TargetProtein Table. This can be in numeric or character vector format}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from the
 TargetProteinGroupsTargetPeptideGroups table or a character string
 specifying anSQL query
}
\description{
get the peptideID's from (a set of) proteinGroupIDs
}
\note{
to get the proteinpeptidelink (table =
 "TargetProteinGroupsTargetPeptideGroups"). In goes "ProteinGroupID" from
 the table "TargetProteins" (Note: it's possible to use a c(,,,) to get
 the result for a number of proteins at the same time). The result is a
 list of numbers which are the "TargetProteinGroupsProteinGroupID" in the
 "TargetPeptideGroups" table
 @export
}
