% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetPeptideTable}
\alias{dbGetPeptideTable}
\title{get the paptide table belonging defined by PeptideIDs ot proteinGroupIDs}
\usage{
dbGetPeptideTable(
  db,
  peptideIDs = NA,
  proteinGroupIDs = NA,
  columns = NA,
  masterProtein = TRUE,
  sortorder = NA,
  SQL = FALSE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{peptideIDs}{the peptideIDs to be retrieved. This can be in numeric or
character vector format OR the output from the dbGetPeptideIDs function
(a data.frame with column "TargetPeptideGroupsPeptideGroupID")}

\item{proteinGroupIDs}{the proteinGroupIDs usually come from the
TargetProtein Table. This can be in numeric or character vector format.
Note: if this parameter is not NA, then peptideIDs will be ignored. This
makes it possible to retrieve the peptides belonging to a protein w/o first
having to retrieve toe Peptide ID's}

\item{columns}{allows the selection of columns to take from the table,
default = NA (all columns)}

\item{masterProtein}{use the IsMasterProtein column to be zero,
default == TRUE. If more advanced filtering is needed, use db_getTable()}

\item{sortorder}{allows for sorting of the selected columns,
default = NA, (no sorting). Other valid values are a single character
string ("ASC" or "DESC") or a character vector of the same length as the
columnNames vector containing a series of "ASC" or "DESC"}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from the peptide table or
 a character string specifying a SQL query
}
\description{
get the paptide table belonging defined by PeptideIDs ot proteinGroupIDs
}
