% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetProteinAnnotationGroupIDs}
\alias{dbGetProteinAnnotationGroupIDs}
\title{Function to get the functional group annotation group ID's for proteins.
 This function does essentially the reverse of
 \code{\link{dbGetAnnotatedProteins}}. The output of this function can serve
 as the input for \code{\link{dbGetAnnotationGroups}}}
\usage{
dbGetProteinAnnotationGroupIDs(db, uniqueSequenceIDs, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{uniqueSequenceIDs}{the UniqueSequenceID's (unique protein identifier),
usually coming from protein table}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame or a character vector (SQL)
}
\description{
Function to get the functional group annotation group ID's for proteins.
 This function does essentially the reverse of
 \code{\link{dbGetAnnotatedProteins}}. The output of this function can serve
 as the input for \code{\link{dbGetAnnotationGroups}}
}
