% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetProteinGroupIDs}
\alias{dbGetProteinGroupIDs}
\title{Retrieve the ProteinGroupID's of proteins via their UniqueSequenceID's}
\usage{
dbGetProteinGroupIDs(db, proteinUniqueIDs, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{proteinUniqueIDs}{the UniqueSequenceID's for which the proteinGroupID's
are to be retrieved. Usually these UniqueSequenceID's will come from a
protein table. Please note that a 'regular' bit64::as.integer64 vector may
fail due to conversion issues. It is better to pass this type of vector as
a character vector}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame#'}
}
\value{
a data.frame or a character vector (SQL)
}
\description{
Retrieve the ProteinGroupID's of proteins via their UniqueSequenceID's
}
\note{
the output of this is meant to serve as input for the
 \code{\link{dbGetProteinGroups}} function
}
