% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetProteinIDs}
\alias{dbGetProteinIDs}
\title{Function to get proteinUniqueID's from a (set of) protein groupID's
 (eg from a proteinGroup tables, or dbGetProteinGroupIDs). This allows for
 getting all proteins (also non-master proteins) which together make up
 a protein group. Normally only the master protein is shown in a protein
 table}
\usage{
dbGetProteinIDs(db, proteinGroupIDs, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{proteinGroupIDs}{the protein group(s) for which the UniqueSequenceID's
should be retrieved. This can also be a (collpased) character vector where
the protein groups are separated by ';'}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame#'}
}
\value{
a data.frame or a character vector (SQL)#'
}
\description{
Function to get proteinUniqueID's from a (set of) protein groupID's
 (eg from a proteinGroup tables, or dbGetProteinGroupIDs). This allows for
 getting all proteins (also non-master proteins) which together make up
 a protein group. Normally only the master protein is shown in a protein
 table
}
\note{
every protein in the protein table has a ProteinGroupID & a
 UniqueSequenceID. The UniqueSequenceID is untique to the protein. A protein
 group may contain more than one protein (and thus also more than one
 UniqueSequenceID)
}
