% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetProteins}
\alias{dbGetProteins}
\title{Function to get protein information from the TargetProteins table on the
 basis of their UniqueSequenceID}
\usage{
dbGetProteins(db, UniqueSequenceIDs, columns = NA, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{UniqueSequenceIDs}{specifies from which proteins to get info}

\item{columns}{character vector, specifies which columns to retrieve}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame#'}
}
\value{
a data.frame or a character vector (SQL)
}
\description{
Function to get protein information from the TargetProteins table on the
 basis of their UniqueSequenceID
}
