% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetProteinTable}
\alias{dbGetProteinTable}
\title{get the protein table from a .pdResult file
(essentially a wrapper around db_getTable())}
\usage{
dbGetProteinTable(
  db,
  columnNames = NA,
  masterProtein = TRUE,
  sortOrder = NA,
  SQL = FALSE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{columnNames}{allows the selection of columns to take from the table,
default = NA (all columns)}

\item{masterProtein}{use the IsMasterProtein column to be zero,
default == TRUE. If more advanced filtering is needed, use db_getTable()}

\item{sortOrder}{allows for sorting of the selected columns,
default = NA, (no sorting). Other valid values are a single character
string ("ASC" or "DESC") or a character vector of the same length as the
columnNames vector containing a series of "ASC" or "DESC"}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from the protein table or
 a character string specifying a SQL query
}
\description{
get the protein table from a .pdResult file
(essentially a wrapper around db_getTable())
}
