% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetProteinUniqueSequenceIDs}
\alias{dbGetProteinUniqueSequenceIDs}
\title{Function to retrieve the UniqueSequenceID's based on the accession field of
 the proteinTable. Essentially a wrapper for
 \code{\link{dbGetProteinFiltered}}}
\usage{
dbGetProteinUniqueSequenceIDs(db, accession = NA, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{accession}{accession(s) of the proteins}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame or a character vector (SQL)
}
\description{
Function to retrieve the UniqueSequenceID's based on the accession field of
 the proteinTable. Essentially a wrapper for
 \code{\link{dbGetProteinFiltered}}
}
