% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetPsmIDs}
\alias{dbGetPsmIDs}
\title{get the PsmID's from (a set of) PeptideGroupIDs}
\usage{
dbGetPsmIDs(db, peptideGroupIDs, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{peptideGroupIDs}{the PeptideGroupIDs usually come from the
TargetPeptideGroups Table. This can be in numeric or character vector format}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from the
 TargetPsmsTargetPeptideGroups  table or a character string specifying
 an SQL query
}
\description{
get the PsmID's from (a set of) PeptideGroupIDs
}
