% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetPsmTable}
\alias{dbGetPsmTable}
\title{get the PSM table belonging to the PsmIDs}
\usage{
dbGetPsmTable(
  db,
  psmIDs = NA,
  peptideGroupIDs = NA,
  columns = NA,
  masterProtein = TRUE,
  sortorder = NA,
  filtering = "MasterProteinAccessions IS NOT NULL",
  SQL = FALSE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{psmIDs}{the PsmIDs to be retrieved. This can be in numeric or
character vector format OR the output from the dbGetPsmIDs function
(a data.frame with column "TargetPsmsPeptideID")}

\item{peptideGroupIDs}{the PeptideGroupIDs usually come from the
TargetPeptideGroups Table. This can be in numeric or character vector format
Note: if this parameter is not NA, then psmIDs will be ignored. This
makes it possible to retrieve the psm info belonging to a peptide w/o first
having to retrieve toe psm ID's}

\item{columns}{allows the selection of columns to take from the table,
default = NA (all columns)}

\item{masterProtein}{use the IsMasterProtein column to be zero,
default == TRUE. If more advanced filtering is needed, use db_getTable()}

\item{sortorder}{allows for sorting of the selected columns,
default = NA, (no sorting). Other valid values are a single character
string ("ASC" or "DESC") or a character vector of the same length as the
columnNames vector containing a series of "ASC" or "DESC"}

\item{filtering}{allows for " WHERE <expression>" additions to the SQL
statement default = " " (no filtering). Note: always put a space (" ")
before any statement. If NA then no filtering is applied. Note that
filtering is only used when the argument PsmIDs is not NA}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from the peptide table or
 a character string specifying a SQL query
}
\description{
get the PSM table belonging to the PsmIDs
}
