% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetQuanSpectrumIDs}
\alias{dbGetQuanSpectrumIDs}
\title{get the SpectrumID's from (a set of) PeptideIDs}
\usage{
dbGetQuanSpectrumIDs(db, peptideIDs, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{peptideIDs}{the PeptideIDs usually come from the
TargetPsms Table. This can be in numeric or character vector format}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from the
 TargetPsmsQuanSpectrumInfo table or a character string specifying
 an SQL query
}
\description{
get the SpectrumID's from (a set of) PeptideIDs
}
