% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetQuanSpectrumInfoTable}
\alias{dbGetQuanSpectrumInfoTable}
\title{get the QuanSpectrumInfo table belonging to the SpectrumID's}
\usage{
dbGetQuanSpectrumInfoTable(
  db,
  SpectrumIDs = NA,
  columnNames = NA,
  masterProtein = TRUE,
  sortOrder = NA,
  SQL = FALSE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{SpectrumIDs}{the SpectrumID's to be retrieved. This can be in numeric
or character vector format OR the output from the dbGetQuanSpectrumIDs
function (a data.frame with column "QuanSpectrumInfoSpectrumID")}

\item{columnNames}{allows the selection of columns to take from the table,
default = NA (all columns)}

\item{masterProtein}{use the IsMasterProtein column to be zero,
default == TRUE. If more advanced filtering is needed, use db_getTable()}

\item{sortOrder}{allows for sorting of the selected columns,
default = NA, (no sorting). Other valid values are a single character
string ("ASC" or "DESC") or a character vector of the same length as the
columnNames vector containing a series of "ASC" or "DESC"}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from the QuanSpectrumInfo
 table or a character string specifying a SQL query
}
\description{
get the QuanSpectrumInfo table belonging to the SpectrumID's
}
