% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetTable}
\alias{dbGetTable}
\title{get a table from a .pdResult file}
\usage{
dbGetTable(
  db,
  tableName,
  columnNames = NA,
  filtering = " ",
  sortOrder = NA,
  SQL = FALSE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{tableName}{used to pass on the name of the table containing the data}

\item{columnNames}{allows the selection of columns to take from the table,
default = NA (all columns)}

\item{filtering}{allows for " WHERE <expression>" additions to the SQL
statement default = " " (no filtering). Note: always put a space (" ")
before any statement}

\item{sortOrder}{allows for sorting of the selected columns,
default = NA, (no sorting). Other valid value is a character character
vector of columnNames to be used for sorting string (with "ASC" or "DESC"
if needed)}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from a database table or
 a character string specifying an SQL query
}
\description{
get a table from a .pdResult file
}
