% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbOpen}
\alias{dbOpen}
\title{Wrapper around pool::dbPool(): opens a database}
\usage{
dbOpen(fileName, drv = RSQLite::SQLite(), ...)
}
\arguments{
\item{fileName}{a character vector specifying the name and location
of the database}

\item{drv}{defines database connection type, default = RSQLite::SQLite()}

\item{...}{to pass on additional parameters to pool::dbPool, exmples are
host = "shiny-demo.csa7qlmguqrf.us-east-1.rds.amazonaws.com"
username = "guest"
password = "guest"}
}
\value{
database access 'handle'
}
\description{
Wrapper around pool::dbPool(): opens a database
}
\note{
if no file with the name 'fileName' exists, then it will be created
(but obviously it will be empty, so most further commands will fail)

if fileName == ":memory:" the database will be an in-memory database
}
