% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{determineBlobLengths}
\alias{determineBlobLengths}
\title{determines the length of the blob (raw) type columns in a data.frame from a
 table (possibly from a database)}
\usage{
determineBlobLengths(blobDF, theTable)
}
\arguments{
\item{blobDF}{the result from either db_getBlobs or getBlobs}

\item{theTable}{the table to which blobDF refers, so either the table from 
the database, which was used for db_getBlobs or the table (dataframe) used
for getBlobs}
}
\value{
blobDF with a single column (named 'length') added which contains
 the length (number of bytes) of each blob column
}
\description{
determines the length of the blob (raw) type columns in a data.frame from a
 table (possibly from a database)
}
\note{
internal function
}
