% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{determineBlobTypeRaw}
\alias{determineBlobTypeRaw}
\title{function which converts the blob length to the appropiate (R) datatype}
\usage{
determineBlobTypeRaw(blobLength)
}
\arguments{
\item{blobLength}{currently only 5 (integer) & 9 (numeric) are supported.
Anything else will result in NA}
}
\value{
a string (either "integer" or "numeric") or NA
}
\description{
function which converts the blob length to the appropiate (R) datatype
}
\note{
internal function
}
