% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{determineRawTypes}
\alias{determineRawTypes}
\title{automatically attempts determines the types (numeric or integer) of the
raw vector columns in a data.frame}
\usage{
determineRawTypes(df, allowSpecials = TRUE)
}
\arguments{
\item{df}{data.frame containing ONLY blob/rawVector columns}

\item{allowSpecials}{allow for the 'special' conversions of columns with
the names specified om the specials data.frame.
Default = TRUE. Note: it is best to force the type
if this is set to FALSE}
}
\value{
a list of the raw types
}
\description{
automatically attempts determines the types (numeric or integer) of the
raw vector columns in a data.frame
}
