% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{dfTransformRaws}
\alias{dfTransformRaws}
\title{df_transform_raws(): converts raw columns in a data.frame to the correct
data types}
\usage{
dfTransformRaws(df, forceBlob = NA, allowSpecials = TRUE)
}
\arguments{
\item{df}{data.frame coming from a table from a Proteome Discoverer
database (eg .pdResult files)}

\item{forceBlob}{must be data.frame with 3 columns: name (columnName),
what (type)& minimumSize (number of values in a cell)
default = NA. forceBlob exists to override automatic
conversion (which has a potential for mistakes when
determining the type of a rawVector in a columnVector).
If 'what' in the forceBlob data.frame = NA, then the
columnVector will not be converted, but returned as it
is!}

\item{allowSpecials}{allow for the 'special' cpnversions of columns with
the names specified om the specials data.frame.
Default = TRUE. Note: it is best to use forceBlob
with 'what' set to NA to prevent conversion problems}
}
\value{
data.frame with all raw vector ('blob') columns converted to more
        more regular data types
}
\description{
df_transform_raws(): converts raw columns in a data.frame to the correct
data types
}
\note{
the tables/data.frame's coming from a Proteome Discoverer database
(eg .pdResult files) have columns of the type raw vecotr (blob).
These can be converted automatically or semi-automatically by this function

If there are no raw vector columns, then this function has no use and
may even trigger errors/warnings

This function can only do integer & numeric blob columns (and the
specials) at the moment

some raw vector columns are actually two (or possibly more) columns
in one. In those cases each element/cell of the column is two (or more)
values. This function splits these columns into two seperate ones.
}
