% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{getAcquistionDate}
\alias{getAcquistionDate}
\title{function to retrieve the acquisition date of the files used to generate the
 pdResult file}
\usage{
getAcquistionDate(db)
}
\arguments{
\item{db}{database access 'handle'}
}
\value{
one or more POSIXct/POSIXt object(S)
}
\description{
function to retrieve the acquisition date of the files used to generate the
 pdResult file
}
\note{
this function is essentially a wrapper around 
 \code{\link{getAcquistionDateTime}}
}
