% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{getAcquistionDateTime}
\alias{getAcquistionDateTime}
\title{function to retrieve the acquisition date & time of the files used to
 generate the pdResult file}
\usage{
getAcquistionDateTime(
  db,
  useAmPm = TRUE,
  format = ifelse(useAmPm, "\%m/\%d/\%Y \%I:\%M:\%S \%p", "\%m/\%d/\%Y \%H:\%M:\%S \%p")
)
}
\arguments{
\item{db}{database access 'handle'}

\item{useAmPm}{logical, influences what default format is used. Ignored if a
format is specified}

\item{format}{character vector specifying the format of the resulting
POSIXct/POSIXt object. See \code{\link[base]{strptime}} for more info}
}
\value{
one or more POSIXct/POSIXt object(S)
}
\description{
function to retrieve the acquisition date & time of the files used to
 generate the pdResult file
}
\note{
this function is essentially a wrapper around 
 \code{\link[proteinDiscover]{studyDefinitionFileSets}}
}
