% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{getBlobs}
\alias{getBlobs}
\title{detemines which columns in a table are of the blob (raw) type}
\usage{
getBlobs(theTable)
}
\arguments{
\item{theTable}{the table containing the data}
}
\value{
a data.frame with two columns: name = colum name) and type (which
 should always be 'blob')
}
\description{
detemines which columns in a table are of the blob (raw) type
}
\note{
meant for use in debugging problems
}
