% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{getPeptideInfoRaw}
\alias{getPeptideInfoRaw}
\title{get peptide information from the peptide table from a pdResult file based on
 the provided proteinAccession (uniprot) codes. Raw columns are not
 "translated"}
\usage{
getPeptideInfoRaw(
  db,
  columnNames = "AbundancesNormalized",
  proteinAccessions = knockOutProteins()$Accession
)
}
\arguments{
\item{db}{database access 'handle'}

\item{columnNames}{allows the selection of columns to take from the table. 
The columns: PeptideGroupID, Sequence, Modifications, QuanInfo are
automatically included. Default column to be retrieved is
AbundancesNormalized}

\item{proteinAccessions}{defines from which protein(s) info will be retrieved
(character vector)}
}
\value{
a named list of data.frames (the names are the proteinAccessions)
}
\description{
get peptide information from the peptide table from a pdResult file based on
 the provided proteinAccession (uniprot) codes. Raw columns are not
 "translated"
}
