% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{getProteinInfo}
\alias{getProteinInfo}
\title{get protein info (with translation of columns) from a list of protein
 Accessions (uniprot code). Essentially this is a wrapper function for
 \code{\link{getProteinInfoRaw}}}
\usage{
getProteinInfo(
  db,
  columnNames = c("Accession", "ProteinGroupIDs", "AbundancesNormalized",
    "AbundanceRatios", "AbundanceRatioPValue", "AbundanceRatioAdjPValue"),
  proteinAccessions = knockOutProteins()$Accession,
  sortOrder = "Accession"
)
}
\arguments{
\item{db}{database access 'handle'}

\item{columnNames}{allows the selection of columns to take from the table}

\item{proteinAccessions}{defines which protein(s) info will be retrieved
(character vector)}

\item{sortOrder}{allows for sorting of the resulting data.frame by on of it's
columns (default = "Accession")}
}
\value{
a data.frame containing requested data from the protein table after
 "translation" of the raw columns
}
\description{
get protein info (with translation of columns) from a list of protein
 Accessions (uniprot code). Essentially this is a wrapper function for
 \code{\link{getProteinInfoRaw}}
}
\note{
this function uses the default
 \code{\link[proteinDiscover]{getProteinInfoRaw}} function. If more control
 over the "translation" of raw columns is needed, then use
 \code{\link{getProteinInfoRaw}} and do the translation manually
}
