% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{getProteinInfoRaw}
\alias{getProteinInfoRaw}
\title{get protein info (without translation of columns) from a list of protein
 Accessions (uniprot code). Essentially this is a wrapper function for
 \code{\link[proteinDiscover]{dbGetTable}}}
\usage{
getProteinInfoRaw(
  db,
  columnNames = c("Accession", "ProteinGroupIDs", "AbundancesNormalized",
    "AbundanceRatios", "AbundanceRatioPValue", "AbundanceRatioAdjPValue"),
  proteinAccessions = knockOutProteins()$Accession,
  sortOrder = "Accession",
  SQL = FALSE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{columnNames}{allows the selection of columns to take from the table}

\item{proteinAccessions}{defines which protein(s) info will be retrieved
(character vector)}

\item{sortOrder}{allows for sorting of the resulting data.frame by on of it's
columns (default = "Accession")}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame (for debugging purposes)}
}
\value{
a data.frame containing requested data from the protein table or
 a character string specifying an SQL query
}
\description{
get protein info (without translation of columns) from a list of protein
 Accessions (uniprot code). Essentially this is a wrapper function for
 \code{\link[proteinDiscover]{dbGetTable}}
}
