% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{isClass}
\alias{isClass}
\title{internal helper function to determine if object == whichClass or a descendant of
 whichClass}
\usage{
isClass(object, whichClass)
}
\arguments{
\item{object}{a data object of some class}

\item{whichClass}{character string: class name to be tested}
}
\value{
TRUE or FALSE
}
\description{
internal helper function to determine if object == whichClass or a descendant of
 whichClass
}
