% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{isMasterProtein}
\alias{isMasterProtein}
\title{function for 'translation' of the isMasterProtein values (0..4) in the
 proteinTable to words (like in Proteome Discoverer).}
\usage{
isMasterProtein(info)
}
\arguments{
\item{info}{integer vector to be 'translated'}
}
\value{
character vector (the translation)
}
\description{
function for 'translation' of the isMasterProtein values (0..4) in the
 proteinTable to words (like in Proteome Discoverer).
}
