% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{knockOutProteins}
\alias{knockOutProteins}
\title{helper function to generate the a data.frame of proteins info for other
 functions}
\usage{
knockOutProteins()
}
\value{
a data.frame with three columns: short (character vector),
 Accession (character vector, uniprot "style") and knockout (logical)
}
\description{
helper function to generate the a data.frame of proteins info for other
 functions
}
\note{
even though it's called knockOutProteins, 2 of the proteins are
 not knock out proteins.
}
