% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{MSfileInfo}
\alias{MSfileInfo}
\title{get the table with info on the files used in the search from the database}
\usage{
MSfileInfo(db, type = "XcaliburRawfile", dates = thermo.date, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{type}{allows for selection of the FileTypes
default = "XCaliburRawFile"}

\item{dates}{allows transformation of the date/time strings from te database
to be transformed into proper data/time fields. Default function used is
thermo.date. If no transformation is required, use na.date}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
data.frame
}
\description{
get the table with info on the files used in the search from the database
}
