% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{na.date}
\alias{na.date}
\title{fake converter for times when no conversion is wanted/needed}
\usage{
na.date(theDate)
}
\arguments{
\item{theDate}{character string
(can be vectorized)}
}
\value{
theDate (original character string)
}
\description{
fake converter for times when no conversion is wanted/needed
}
