% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workFlows.R
\name{nodeTable}
\alias{nodeTable}
\title{function to display an overview table of the processing/consensus workflows
 in the nodeInfo coming out of the workflowInfo function}
\usage{
nodeTable(nodeInfo)
}
\arguments{
\item{nodeInfo}{either the processing or consensus part of the nodeInfo}
}
\description{
function to display an overview table of the processing/consensus workflows
 in the nodeInfo coming out of the workflowInfo function
}
\note{
an example of it's use: (workflowInfo(db))$nodeInfo$Consensus
 %>% nodeTable()
}
