% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{pQuanInfo}
\alias{pQuanInfo}
\title{function for translation of the QuanInfos values in the psms & peptide
 tables to words (like in Proteome Discoverer).}
\usage{
pQuanInfo(info)
}
\arguments{
\item{info}{integer vector to be 'translated'}
}
\value{
character vector (the translation)
}
\description{
function for translation of the QuanInfos values in the psms & peptide
 tables to words (like in Proteome Discoverer).
}
