% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{proteinIDTypes}
\alias{proteinIDTypes}
\title{get the names of the identification types (sequest HT etc) used in the
 database}
\usage{
proteinIDTypes(db, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
data.frame with a single column: "GroupName"
}
\description{
get the names of the identification types (sequest HT etc) used in the
 database
}
