% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{psmAmbiguity}
\alias{psmAmbiguity}
\title{function for 'translation' of the psmAmbiguity values (1..5) in the psmTable
 to words (like in Proteome Discoverer). <...> --> means not encountered/
 undefined/no inference}
\usage{
psmAmbiguity(info)
}
\arguments{
\item{info}{integer vector to be 'translated'}
}
\value{
character vector (the translation)
}
\description{
function for 'translation' of the psmAmbiguity values (1..5) in the psmTable
 to words (like in Proteome Discoverer). <...> --> means not encountered/
 undefined/no inference
}
