% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{quanInfo}
\alias{quanInfo}
\title{function for 'translation' of the QuanInfo values in the QuanSpectrumInfo
 table to words (like in Proteome Discoverer).}
\usage{
quanInfo(info)
}
\arguments{
\item{info}{integer vector to be 'translated'}
}
\value{
character vector (the translation)
}
\description{
function for 'translation' of the QuanInfo values in the QuanSpectrumInfo
 table to words (like in Proteome Discoverer).
}
