% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDefinition.R
\name{replacementStrings}
\alias{replacementStrings}
\title{function that generates the default data.frame for the function
df_replace().}
\usage{
replacementStrings()
}
\value{
a data.frame with columns:
 value, replacement and singleChar
}
\description{
function that generates the default data.frame for the function
df_replace().
}
\note{
value is the string to be searched, replacement is what it needs
 to be replaced with. singleChar sets whether the replacement should only
 take place when dealing with single character strings. This is because
 single character strings sometimes 'act' different when rendering
 markdown documents in HTML
}
