% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectrum.precursor.header}
\alias{spectrum.precursor.header}
\title{gets the info in the list object coming from the function
'transformSpectrumRaw': spectrum parent header}
\usage{
spectrum.precursor.header(spectrum)
}
\arguments{
\item{spectrum}{list object containing info on a spectrum}
}
\value{
data.frame
}
\description{
gets the info in the list object coming from the function
'transformSpectrumRaw': spectrum parent header
}
