% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectrum.precursor.info}
\alias{spectrum.precursor.info}
\title{gets the info in the list object coming from the function
'transformSpectrumRaw': spectrum parent monoisotopic peak}
\usage{
spectrum.precursor.info(spectrum, measured = TRUE)
}
\arguments{
\item{spectrum}{list object containing info on a spectrum}

\item{measured}{logical vector, if TRUE then the measured data is returned}
}
\value{
data.frame
}
\description{
gets the info in the list object coming from the function
'transformSpectrumRaw': spectrum parent monoisotopic peak
}
