% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectrum.precursor.scanEvent}
\alias{spectrum.precursor.scanEvent}
\title{gets the info in the list object coming from the function
'transformSpectrumRaw': spectrum parent scan event}
\usage{
spectrum.precursor.scanEvent(spectrum, returnRaw = FALSE)
}
\arguments{
\item{spectrum}{list object containing info on a spectrum}

\item{returnRaw}{logical vector, if TRUE them the data is returned as a list. if FALSE (default)
then a data.frame of all character-type data is returned'}
}
\value{
data.frame or list
}
\description{
gets the info in the list object coming from the function
'transformSpectrumRaw': spectrum parent scan event
}
