% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectrum.profile}
\alias{spectrum.profile}
\title{gets the info in the list object coming from the function
'transformSpectrumRaw': spectrum profile spectrum}
\usage{
spectrum.profile(spectrum)
}
\arguments{
\item{spectrum}{list object containing info on a spectrum}
}
\value{
NULL or NA
}
\description{
gets the info in the list object coming from the function
'transformSpectrumRaw': spectrum profile spectrum
}
\note{
this is a convenience function, type of data not observed
}
