% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectrum.scanEvent}
\alias{spectrum.scanEvent}
\title{gets the info in the list object coming from the function
'transformSpectrumRaw': spectrum scan event}
\usage{
spectrum.scanEvent(spectrum)
}
\arguments{
\item{spectrum}{list object containing info on a spectrum}
}
\value{
data.frame
}
\description{
gets the info in the list object coming from the function
'transformSpectrumRaw': spectrum scan event
}
