% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDefinition.R
\name{studyDefinitionExtensions}
\alias{studyDefinitionExtensions}
\title{function that extracts information on isotope corrections (if available)}
\usage{
studyDefinitionExtensions(analysisDef, correctXML = c("utf-16", "utf-8"))
}
\arguments{
\item{analysisDef}{generated by the analysisDefinition function}

\item{correctXML}{can only have two different values: NA or a two element
character vector c("utf-16","utf-8"). During the research into the
method descriptions in the XML object it was noticed that the
XML::xmlToList gave an error Document labelled UTF-16 but has
UTF-8 content. This was solved by replacing the 'utf-16' string by 'utf-8'
string in the XML object. This may be a country specific issue, so the
function allows setting this parameter to NA will not do the replacement.}
}
\value{
NA or a list of two data.frame objects
}
\description{
function that extracts information on isotope corrections (if available)
}
