% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDefinition.R
\name{studyDefinitionExtensionSettings}
\alias{studyDefinitionExtensionSettings}
\title{function to extract sample/factor/ratio/replicate information.}
\usage{
studyDefinitionExtensionSettings(analysisDef)
}
\arguments{
\item{analysisDef}{generated by the analysisDefinition function}
}
\value{
a lits of 4 elements:
\enumerate{
  \item StudyVariablesForGrouping : a data.frame of factors used
  \item StudyVariablesForSorting : a data.frame of sorting specification for
   the factors
  \item QuanRatios : a list object of all ratios. Each ratio has the
   following elements: RatioTable (specifying numerator/denominator),
   RatioString (for easy info printing), NumeratorSamples &
   DenominatorSamples specifying which samples are in the numerator and
   denominator and finally Replicates which contains info on replicates.
  \item XML : the actual from which the information comes. This was included
   because the exact specification for all possible cases is not (yet) known
}
}
\description{
function to extract sample/factor/ratio/replicate information.
}
\note{
So far, this function has not been tested for all possible cases/
 scenarios.
}
