% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDefinition.R
\name{studyDefinitionFactors}
\alias{studyDefinitionFactors}
\title{function that extracts the factors used in the study to generate the
 .pdResult file. The result contains some internal info in the form
 of columns named id (identifiers).}
\usage{
studyDefinitionFactors(analysisDef)
}
\arguments{
\item{analysisDef}{generated by the analysisDefinition function}
}
\value{
data.frame with the info
}
\description{
function that extracts the factors used in the study to generate the
 .pdResult file. The result contains some internal info in the form
 of columns named id (identifiers).
}
