% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDefinition.R
\name{studyDefinitionFileSets}
\alias{studyDefinitionFileSets}
\title{function that extracts file information on the original .raw files
 used to generate the .pdResult file. Information includes the
 original file name, location & size. It also contains some internal info in
 the form of columns named id (identifiers).}
\usage{
studyDefinitionFileSets(analysisDef, splitFileSize = TRUE, joinedTables = TRUE)
}
\arguments{
\item{analysisDef}{generated by the analysisDefinition function}

\item{splitFileSize}{boolean (default: TRUE), specifies if the FileSize
column should be split into the actual file size (still a character vector)
and the file size format}

\item{joinedTables}{boolean (default: TRUE), specifies if all info should be
put in a single data.frame. If FALSE it will generate a list of two
data.frame objects; this might be useful in some scenarios}
}
\value{
data.frame or list of two data.frame objects
}
\description{
function that extracts file information on the original .raw files
 used to generate the .pdResult file. Information includes the
 original file name, location & size. It also contains some internal info in
 the form of columns named id (identifiers).
}
