% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDefinition.R
\name{studyDefinitionQuanMethods}
\alias{studyDefinitionQuanMethods}
\title{function that extracts quantification method information if a quantification
 method was used to generate the .pdResult file}
\usage{
studyDefinitionQuanMethods(analysisDef)
}
\arguments{
\item{analysisDef}{generated by the analysisDefinition function}
}
\value{
A list of two data.frame objects. The first one will contain the
 name, description, etc. The second one will specify the names of the labels
 used. The result will be NA in the case that no quantification method was
 used.
}
\description{
function that extracts quantification method information if a quantification
 method was used to generate the .pdResult file
}
