% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyDefinition.R
\name{studyDefinitionSamples}
\alias{studyDefinitionSamples}
\title{function that extracts sample information. The information seems to be a bit
 redundant, as the info is also seen in other tables.}
\usage{
studyDefinitionSamples(analysisDef)
}
\arguments{
\item{analysisDef}{generated by the analysisDefinition function}
}
\value{
a data.frame
}
\description{
function that extracts sample information. The information seems to be a bit
 redundant, as the info is also seen in other tables.
}
