% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{tableName}
\alias{tableName}
\title{helper function to prevent having to remember the somewhat long names of
the most used tables}
\usage{
tableName(whichTable = "proteins")
}
\arguments{
\item{whichTable}{can be either "proteins","peptides","psms" or "consensus"
character do not need to be lower or upper case (all are converted to upper
case). If another string is used as a parameter, the function will return
NA}
}
\value{
a string containing the protein discoverer table name corresponding
 to the parameter whichTable
}
\description{
helper function to prevent having to remember the somewhat long names of
the most used tables
}
