% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{tmt10Channels}
\alias{tmt10Channels}
\title{helper function to generate the a data.frame of TMT knockout strain (TKO)
 info for other functions. This function generates a data.frame based on
 the 10-plex TMT TKO knockout (this was the original TMT-knockout-digest
 available)}
\usage{
tmt10Channels()
}
\value{
a data.frame with four columns: all are character vectors
}
\description{
helper function to generate the a data.frame of TMT knockout strain (TKO)
 info for other functions. This function generates a data.frame based on
 the 10-plex TMT TKO knockout (this was the original TMT-knockout-digest
 available)
}
\note{
the rows define the order of the abundance (etc) columns in the
 protein, peptide and psms table in a pdResult file. The order is
 alphabetical in protein & peptide tables, but not in the psms tables: there
 it is based based on the order of the isotopes

psmsChannels & isotopeChannels columns match each other
}
