% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{totalSearchTime}
\alias{totalSearchTime}
\title{get the total search time  from the database}
\usage{
totalSearchTime(db)
}
\arguments{
\item{db}{database access 'handle'}
}
\value{
numeric: search time in seconds
}
\description{
get the total search time  from the database
}
