% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{transformSpectrumRaw}
\alias{transformSpectrumRaw}
\title{transforms a spectrum from the table 'MassSpectrumItems' into a R compatible
 list}
\usage{
transformSpectrumRaw(spectrumObject)
}
\arguments{
\item{spectrumObject}{must be of class 'raw'}
}
\value{
a list object containing info on the spectrum (object). This list
 object can be further translated via the function 'translateSpectrumInfo'
}
\description{
transforms a spectrum from the table 'MassSpectrumItems' into a R compatible
 list
}
\note{
this functions writes a temporary file tp disk, which is unzipped, read
 and deleted again
}
