% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{translateInfo}
\alias{translateInfo}
\title{internal function: function factory for 'translating' numbers to their
 corresponding strings/messages}
\usage{
translateInfo(translation, setNAZero = TRUE, zeroIsFirst = TRUE)
}
\arguments{
\item{translation}{character vector specifying the strings/messages}

\item{setNAZero}{boolean specifying whether NA values should be converted to
zero}

\item{zeroIsFirst}{boolean specifying whether the first element is zero or 1
(for dealing with zero values). If TRUE then all values are increased by 1}
}
\value{
a function which converts a (vector of) number(s) to the specified
 strings/messages
}
\description{
internal function: function factory for 'translating' numbers to their
 corresponding strings/messages
}
