% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{whichRaw}
\alias{whichRaw}
\title{Determine the type of raw vector a rawVector is}
\usage{
whichRaw(rawVector, number = FALSE)
}
\arguments{
\item{rawVector}{A vector of type 'raw' (blob)}

\item{number}{boolean to specifiy if the return value should be
be character (FALSE) or numeric (TRUE). Default = TRUE}
}
\value{
character string ("integer", "numeric") or a numeric value (5 or 9)
}
\description{
Determine the type of raw vector a rawVector is
}
\note{
this method has a weakness, if raw vector length = 45 or similar,
      then the type returned will be integer, even though it may very well
      be numeric  45 %% 5 = 0. but 45 %% 9 = 0 too!
}
