% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{whichRawList}
\alias{whichRawList}
\title{Determine the type of a list of raw vectors}
\usage{
whichRawList(blobList, naValue = NA)
}
\arguments{
\item{blobList}{A list of raw vectors, usually the column of a data.frame}

\item{naValue}{A default value to be used, when it the type cannot be
determined (when a column contains only NA's)}
}
\value{
a list of (type = , number = ) where type is the type of raw vector
        and number is the number of items per cell
}
\description{
Determine the type of a list of raw vectors
}
\note{
Some of the data in the tables that come out of the database system
in # proteome discoverer are in the <blob> raw Vector format. So some
columns are lists of raw Vectors. To convert them, one has to know the type
of raw vector (integer/numeric). Because some elements in the 'blob' list
are NA, this function seeks the first non-NA element and tries to determine
it's type. If it the column contains only NA's then the naValue will be
returned (default = NA). This was included to force a certain type
even if the column (temporarily?) contains no data

only properly works with "integer" or "numeric" raw vectors
}
