% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{whichRawListSpecial}
\alias{whichRawListSpecial}
\title{Determine the type of a list of raw vectors that is a columnSpecial}
\usage{
whichRawListSpecial(blobList, naValue = NA, specialName)
}
\arguments{
\item{blobList}{A list of raw vectors, usually the column of a data.frame}

\item{naValue}{A default value to be used, when it the type cannot be
determined (when a column contains only NA's)}

\item{specialName}{name of the column which is a special}
}
\value{
a list of (type = , number = ) where type is the type of raw vector
        and number is the number of items per cell
}
\description{
Determine the type of a list of raw vectors that is a columnSpecial
}
\note{
the type is always either NA or "special". This function is actually
more for determining the number of items per cell, which is a list
of boolean values
}
