% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workFlows.R
\name{workflowInfo}
\alias{workflowInfo}
\title{function to get the workflow information from a .pdResult file}
\usage{
workflowInfo(db, workflowsTable = "WorkFlows", returnNodeData = TRUE)
}
\arguments{
\item{db}{database access 'handle' pointing to a .pdResult file}

\item{workflowsTable}{name of the table containing the info. Default is
'WorkFlows'}

\item{returnNodeData}{if TRUE then the node parameters are included in the
returned data}
}
\value{
either a single data.frame containing basic info on the workflows or
 (if returnNodeData is TRUE) a list of the data.frame with the second list
 element containing information on the nodes that make up the
  processing & the consensus workflows (in xmlToList result
 format). This second element (called nodeInfo) is used in additional
 functions to show/display the processing/consensus workflows.
}
\description{
function to get the workflow information from a .pdResult file
}
