% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{columnSpecials}
\alias{columnSpecials}
\title{Specials are not numeric or integer, but have chunks of a certain size
All encountered in Proteome Discoverer are actually booleans with a value
0 (FALSE), 1 (TRUE) or NA}
\usage{
columnSpecials()
}
\value{
data.frame with columns 'names' and 'size'
}
\description{
Specials are not numeric or integer, but have chunks of a certain size
All encountered in Proteome Discoverer are actually booleans with a value
0 (FALSE), 1 (TRUE) or NA
}
\note{
each chunk consists of two bytes, first one is logical (boolean):
zero = FALSE, otherwise TRUE. Second byte = also logical: determines if
value is NA (1) or not (0)
}
