% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{dbGetMassSpectrumItems}
\alias{dbGetMassSpectrumItems}
\title{get the MassSpectrumItems info from (a set of) PeptideID's}
\usage{
dbGetMassSpectrumItems(db, dbDetail = NA, peptideID, SQL = FALSE)
}
\arguments{
\item{db}{database access 'handle' (to the .pdResult file)}

\item{dbDetail}{database access 'handle' to the details file (.pdResultDetails). This is needed
for at least Proteome Discover 3.1, since the "MassSpectrumItems" table is located in a
different file than the e.g. the psm table. Note that if the 'SQL' parameter is set to TRUE, the
function will only return the last SQL query (querying the .pdResultDetails table).}

\item{peptideID}{the PeptideID's usually come from the
PSMS table Table. This can be in numeric/character/data.frame format}

\item{SQL}{allows the function to return the SQL query statement in stead of
a data.frame}
}
\value{
a data.frame containing requested data from the MassSpectrumItems
 table or a character string specifying an SQL query
}
\description{
get the MassSpectrumItems info from (a set of) PeptideID's
}
