% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRaw.R
\name{dfTransformRaws}
\alias{dfTransformRaws}
\title{df_transform_raws(): converts raw columns in a data.frame to the correct
data types}
\usage{
dfTransformRaws(
  df,
  blobDF = NA,
  minimumNumber = 1,
  numberOfGroups = minimumNumber,
  ratioNumberOfGroups = numberOfGroups - 1,
  specials = TRUE
)
}
\arguments{
\item{df}{data.frame coming from a table from a Proteome Discoverer
database (eg .pdResult files)}

\item{blobDF}{must be data.frame with 3 columns: name (columnName),
what (type) & minimumSize (number of values in a cell) default = NA.
If 'what' in the data.frame = NA, then the columnVector will not be
converted, but returned as it is}
}
\value{
data.frame with all raw vector ('blob') columns converted to more
        more regular data types
}
\description{
df_transform_raws(): converts raw columns in a data.frame to the correct
data types
}
\note{
the tables/data.frame's coming from a Proteome Discoverer database
(eg .pdResult files) have columns of the type raw vecotr (blob).
These can be converted automatically or semi-automatically by this function

If there are no raw vector columns, then this function has no use and
may even trigger errors/warnings

This function can only do integer & numeric blob columns (and the
specials) at the moment

some raw vector columns are actually two (or possibly more) columns
in one. In those cases each element/cell of the column is two (or more)
values. This function splits these columns into two seperate ones.
}
