% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{getPeptideInfo}
\alias{getPeptideInfo}
\title{get peptide information from the peptide table from a pdResult file based on
 the provided proteinAccession (uniprot) codes. Raw columns are "translated"}
\usage{
getPeptideInfo(
  db,
  columns = "AbundancesNormalized",
  addStandardColumns = TRUE,
  proteinAccessions = knockOutProteins()$Accession,
  removeUnusedQuantInfo = TRUE
)
}
\arguments{
\item{db}{database access 'handle'}

\item{columns}{allows the selection of columns to take from the table. 
The columns: PeptideGroupID, Sequence, Modifications, QuanInfo are
automatically included. Default column to be retrieved is
AbundancesNormalized}

\item{addStandardColumns}{if TRUE then the following columns are added by
default to the columnNames argument: "PeptideGroupID", "Sequence",
"Modifications" & "QuanInfo". Please note that this will give problems if
these columns are also in the columnNames argument. Also: to be able to
use the argument removeUnusedQuantInfo = TRUE, you MUST retrieve the
"QuantInfo" column}

\item{proteinAccessions}{defines from which protein(s) info will be retrieved
(character vector)}

\item{removeUnusedQuantInfo}{default = TRUE. IF TRUE then only peptide info
rows with NA as QuantInfo are kept (the others contain problematic abundance
info or none at all)}
}
\value{
a named list of data.frames (the names are the proteinAccessions)
}
\description{
get peptide information from the peptide table from a pdResult file based on
 the provided proteinAccession (uniprot) codes. Raw columns are "translated"
}
\note{
this function uses the default
 \code{\link[proteinDiscover]{getProteinInfoRaw}} function. If more control
 over the "translation" of raw columns is needed, then use
 \code{\link{getPeptideInfoRaw}} and do the translation manually
}
