% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\docType{data}
\name{IsMasterProtein}
\alias{IsMasterProtein}
\title{for translation of the value (0..4) in the proteinTable to words
(like in Proteome Discoverer). <...> --> means not encountered/undefined/
no inference}
\format{
character vector
}
\usage{
IsMasterProtein
}
\description{
for translation of the value (0..4) in the proteinTable to words
(like in Proteome Discoverer). <...> --> means not encountered/undefined/
no inference
}
\keyword{datasets}
