% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workFlows.R
\name{nodes}
\alias{nodes}
\title{function that takes a (xmlToList type) workflow and returns a list of}
\usage{
nodes(
  Workflow,
  showHidden = FALSE,
  showAdvanced = TRUE,
  showConfiguration = FALSE
)
}
\arguments{
\item{Workflow}{a (xmlToList type) workflow}

\item{showHidden}{if TRUE then rows with hidden = TRUE are included (default:
false)}

\item{showAdvanced}{if TRUE then rows with advanced = TRUE are included
(default: TRUE)}

\item{showConfiguration}{if TRUE then rows with configuration = TRUE are
included (default: FALSE)}
}
\value{
a list of named data.frame objects containing all the parameters/
 settings in the nodes of the workflow
}
\description{
function that takes a (xmlToList type) workflow and returns a list of
}
\note{
an example of it's use: (workflowInfo(db))$nodeInfo$Consensus %>%
 nodes()
}
