% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\docType{data}
\name{PsmAmbiguity}
\alias{PsmAmbiguity}
\title{for translation of the value (1..5) in the psmTable to words
(like in Proteome Discoverer). <...> --> means not encountered/undefined/
no inference}
\format{
character vector
}
\usage{
PsmAmbiguity
}
\description{
for translation of the value (1..5) in the psmTable to words
(like in Proteome Discoverer). <...> --> means not encountered/undefined/
no inference
}
\keyword{datasets}
