% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{system.date}
\alias{system.date}
\title{converts character string date into date/time format
(ymd hms format)}
\usage{
system.date(theDate)
}
\arguments{
\item{theDate}{character string to be converted
(can be vectorized)}
}
\value{
date
}
\description{
converts character string date into date/time format
(ymd hms format)
}
