% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinDatabase.R
\name{thermo.date}
\alias{thermo.date}
\title{converts character string date into date/time format
(mdy hms format)}
\usage{
thermo.date(theDate)
}
\arguments{
\item{theDate}{character string to be converted
(can be vectorized)}
}
\value{
date in mdy hms format
}
\description{
converts character string date into date/time format
(mdy hms format)
}
